cleanup unused requires and refactoring into model/controller/service structure, pokeTimesheets to test JWT authentication
