+++ app.js
@@ -1,9 +1,8 @@
 const express = require("express");
 const fs = require("fs");
 const https = require("https");
-const session = require("express-session");
 const bodyParser = require("body-parser");
-const authService = require("./services/authenticateSrv");
+const passportJwtStrategy = require("./passport");
 
 const app = express();
 const port = 3000;
@@ -19,27 +18,7 @@ app.use(bodyParser.json());
 app.use(express.static(path));
 
 // middleware
-// app.use(
-//   session({
-//     resave: false, // don't save session if unmodified
-//     saveUninitialized: false, // don't create session until something stored
-//     secret: "shhhh, very secret",
-//   }),
-// );
-
-authService();
-
-// Session-persisted message middleware
-// app.use(function(req, res, next) {
-//   var err = req.session.error;
-//   var msg = req.session.success;
-//   delete req.session.error;
-//   delete req.session.success;
-//   res.locals.message = "";
-//   if (err) res.locals.message = "<p class='msg error'>" + err + "</p>";
-//   if (msg) res.locals.message = "<p class='msg success'>" + msg + "</p>";
-//   next();
-// });
+passportJwtStrategy();
 
 app.use("/api/login", login);
 app.use("/api/logout", logout);
