+++ controllers/loginCtrl.js
@@ -1,13 +1,11 @@
-const passport = require("passport");
 const jwt = require("jsonwebtoken");
-const users = require("../models/userMdl");
+const authService = require("../services/authenticateSrv");
 
 function buildToken(payload) {
-  console.log(payload.username);
-  let user = users.find(user => user.username == payload.username);
+  let user = authService(payload);
   // query the db for the given username
   if (!user) return { error: "No user found" };
-  const token = jwt.sign(user, "veryVerySecretKey");
+  const token = jwt.sign(user, "veryVerySecretKey", { expiresIn: "30m" });
 
   return { userId: user.id, token };
 }
